提供多个玩家邮件支持

+++ jeecg-boot/jeecg-boot-module-system/src/main/java/org/jeecg/modules/game/service/impl/GameEmailServiceImpl.java
@@ -85,7 +85,7 @@ public class GameEmailServiceImpl extends ServiceImpl<GameEmailMapper, GameEmail
         }
 
         if (gameEmail.getTargetBodyType() == TARGET_BODY_TYPE) {
-            Integer targetBodyId = gameEmail.getTargetBodyId();
+            Long targetBodyId = gameEmail.getTargetBodyId();
             PlayerRegisterInfo registerInfo = getPlayerRegisterInfo(targetBodyId);
             if (registerInfo == null) {
                 response.setFailure("玩家ID不存在！");
